requirements.txt
torch>=2.0.0,<3.0.0
transformers>=4.35.0
datasets>=2.15.0
evaluate>=0.4.0
rouge_score>=0.1.2
pyyaml>=6.0.1
numpy>=1.24.0
tqdm>=4.65.0